Refactored Chitter app with partial
