+++ Gemfile
@@ -6,6 +6,8 @@ gem 'dm-postgres-adapter'
 gem 'sinatra'
 gem 'bcrypt'
 gem 'sinatra-flash'
+gem 'sinatra-partial'
+gem 'rake'
 
 group :test do
   gem 'rspec'
