+++ app/app.rb
@@ -1,0 +1,10 @@
+ENV["RACK_ENV"] ||= "development"
+
+require 'sinatra/base'
+require 'sinatra/flash'
+require 'sinatra/partial'
+
+require_relative '../data_mapper_setup'
+require_relative 'server'
+require_relative 'controllers/sessions'
+require_relative 'controllers/users'
