+++ app/controllers/sessions.rb
@@ -1,0 +1,32 @@
+class Chitter < Sinatra::Base
+
+  get '/' do
+     redirect to('/home')
+   end
+
+   get '/home' do
+     erb(:home)
+   end
+
+   get '/sessions/new' do
+     erb(:'/sessions/new')
+   end
+
+   post '/sessions' do
+     user = User.authenticate(params[:username], params[:password])
+     if user
+       session[:user_id] = user.id
+       redirect to('/home')
+     else
+       flash.now[:errors] = ['The email or password is incorrect']
+       erb(:'sessions/new')
+     end
+   end
+
+   delete '/sessions' do
+     session[:user_id] = nil
+     flash.keep[:notice] = 'Bye!'
+     redirect to('/home')
+   end
+
+end
