+++ app/controllers/users.rb
@@ -1,0 +1,27 @@
+class Chitter < Sinatra::Base
+
+  get '/users/new' do
+    @user = User.new
+    erb(:'users/new')
+  end
+
+  post '/users' do
+    @user = User.new(username: params[:username],
+                      name: params[:name],
+                      email: params[:email],
+                      password: params[:password],
+                      password_confirmation: params[:password_confirmation])
+    if @user.save
+      session[:user_id] = @user.id
+      redirect to('/home')
+    else
+      flash.now[:errors] = @user.errors.full_messages
+      erb(:'users/new')
+    end
+  end
+
+
+  # start the server if ruby file executed directly
+  run! if app_file == $0
+
+end
