+++ app/models/user.rb
@@ -1,7 +1,7 @@
 require 'data_mapper'
 require 'dm-postgres-adapter'
 require 'bcrypt'
-
+require_relative '../../data_mapper_setup'
 
 class User
 
@@ -32,8 +32,3 @@ class User
     end
   end
 end
-
-# DataMapper.setup(:default, ENV['DATABASE_URL'] || "postgres:///chitter_#{ENV['RACK_ENV']}")
-DataMapper.setup(:default, "postgres://localhost/chitter_test")
-DataMapper.finalize
-DataMapper.auto_upgrade!
