+++ app/server.rb
@@ -1,0 +1,17 @@
+class Chitter < Sinatra::Base
+
+  use Rack::MethodOverride
+  register Sinatra::Flash
+  register Sinatra::Partial
+  enable :sessions
+  set :session_secret, 'super secret'
+  set :partial_template_engine, :erb
+
+  enable :partial_underscores
+
+  helpers do
+    def current_user
+      @current_user ||= User.get(session[:user_id])
+    end
+  end
+end
