+++ app/views/layout.erb
@@ -8,21 +8,8 @@
   <body>
     <h1>Welcome to Chitter</h2>
 
-
-    <% if flash[:errors] && !flash[:errors].empty? %>
-      Please refer to the following errors below:
-      <ul id='errors'>
-        <% flash[:errors].each do |error| %>
-          <li><%= error %></li>
-        <% end %>
-      </ul>
-    <% end %>
-
-    <% if flash[:notice] %>
-      <div id='notice'>
-        <%= flash[:notice] %>
-      </div>
-    <% end %>
+    <%= partial :welcome %>
+    <%= partial :flash %>
 
     <% if current_user %>
       Hello <%= current_user.username %>!
