+++ config.ru
@@ -1,4 +1,4 @@
 require 'rubygems'
-require File.join(File.dirname(__FILE__), 'app.rb')
+require File.join(File.dirname(__FILE__), 'app/app.rb')
 
 run Chitter
