+++ spec/spec_helper.rb
@@ -1,11 +1,11 @@
 # Generated by rspec-sinatra. (2016-10-22 16:34:15 +0100)
+ENV["RACK_ENV"] = "test"
 
 require 'coveralls'
 require 'simplecov'
 
-ENV['RACK_ENV'] = 'test'
 
-require File.join(File.dirname(__FILE__), '..', 'app.rb')
+require File.join(File.dirname(__FILE__), '..', './app/app.rb')
 
 require 'capybara'
 require 'capybara/rspec'
@@ -14,7 +14,7 @@ require 'database_cleaner'
 
 require_relative 'helpers/sessions'
 require './app/models/user'
-require './app'
+require './app/app'
 
 Capybara.app = Chitter
